Assistant checkpoint: Add How to Play instructions modal
Assistant generated file changes:
- client/src/components/game/how-to-play-dialog.tsx: Create a How to Play dialog component, Make dialog controllable by parent
- client/src/pages/game.tsx: Add How to Play button in game UI
Add How to Play button in game UI
Show How to Play dialog automatically for new users
Add How to Play button in game UI
Show How to Play dialog automatically for new users
Add controlled How to Play dialog

---

User prompt:

Users need a way to understand how to play the game because not all of them will be familiar with the rules of this game. Can you think through how we can display in very simple language how to play this game? I think they should be a model that pops up And the instructions should be in very simple language. Please think through this and come up with a plan first.